refactor: PG사 api 호출 로직 트랜잭션 분리
