+++ src/main/kotlin/com/petqua/application/payment/PaymentService.kt
@@ -1,9 +1,9 @@
 package com.petqua.application.payment
 
 import com.petqua.application.order.dto.PayOrderCommand
-import com.petqua.application.payment.infra.PaymentGatewayClient
 import com.petqua.domain.order.OrderRepository
 import com.petqua.domain.order.findByOrderNumberOrThrow
+import com.petqua.domain.payment.tosspayment.TossPayment
 import com.petqua.domain.payment.tosspayment.TossPaymentRepository
 import com.petqua.exception.order.OrderException
 import com.petqua.exception.order.OrderExceptionType
@@ -15,17 +15,17 @@ import org.springframework.transaction.annotation.Transactional
 class PaymentService(
     private val orderRepository: OrderRepository,
     private val paymentRepository: TossPaymentRepository,
-    private val paymentGatewayClient: PaymentGatewayClient,
 ) {
 
-    fun payOrder(command: PayOrderCommand) {
+    @Transactional(readOnly = true)
+    fun validateAmount(command: PayOrderCommand) {
         val order = orderRepository.findByOrderNumberOrThrow(command.orderNumber) {
             OrderException(OrderExceptionType.ORDER_NOT_FOUND)
         }
-
         order.validateAmount(command.amount)
+    }
 
-        val paymentResponse = paymentGatewayClient.confirmPayment(command.toPaymentConfirmRequest())
-        paymentRepository.save(paymentResponse.toPayment())
+    fun save(tossPayment: TossPayment): TossPayment {
+        return paymentRepository.save(tossPayment)
     }
 }
